test: separate credential-home checks from runtime integration tests
