+++ sdk/typescript/TESTING.md
@@ -74,15 +74,18 @@ Windows runs seven separate jobs with `node scripts/run-ci-tests.mjs 1/7`
 (substitute the shard number), each using up to two Bun processes with separate
 reports such as `junit-1-1.xml` and `junit-1-2.xml`. Each file runs once. New files
 are included automatically; stale timing estimates can affect balance, but not coverage.
+Credential-home locking and ACL checks live in `runtime-credentials.test.ts` so
+they can run independently of the plugin, output-directory, and Python checks
+in `runtime.test.ts`.
 The machine-wide Windows policy test still runs separately and serially.
 Windows installs the pnpm version from `packageManager` directly, reuses the
 npm download cache, and caches the resolved pnpm store separately. Cache
 failures do not suppress installation failures. Unix keeps its pinned pnpm
 setup action. Windows package inspection enables npm's native phase timings to
 diagnose installation delays without changing its failure or timeout behavior.
-It also logs npm cache/fetch activity and trials a 16-thread libuv pool for the
-observed unpacking delays. This is a scheduling experiment, not a proven speedup;
-compare hosted install phases and whole-workflow timings before retaining it.
+It also logs npm cache/fetch activity and uses a 16-thread libuv pool for
+filesystem-heavy package extraction. These settings apply only to Windows
+package inspection; the fresh consumer install and its assertions are unchanged.
 
 When forwarding `--test-name-pattern` to a sharded run, also pass Bun's
 `--pass-with-no-tests` if some workers may have no matching tests. Normal CI
